Remove handler callbacks before setting disposed.
This change ensures that you can never observe `disposed` being set to `true` inside `run()`. `removeCallbacks` and the mechanism by which a `Looper` retrieves the head of the message queue are both governed by a lock. When `dispose()` is called on a non-main thread taking this lock is the race which determines cancelation and **not** the boolean.

Include a comment emphasizing that the boolean is tracked solely for the purposes of accurate `isDisposed()` reporting.